requirements.txt
numpy==1.18.2
tqdm==4.43.0
scipy==1.4.1
tensorflow==2.2.0
Pillow==7.1.2